+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.60.7] - 2020-04-30 
+## Fixed
+- <no-ticket> - Busines Rule Component validation added on endDate when one-time event type selected
+              - Hide Sections List when no events available or empty list on calendar
+              - Placeholder on endDate removed when field is disabled
+
 ## [0.60.6] - 2020-03-28  
 ## Fixed
 - CXV1-22314 - Disposition list in config ui and in skylight size adjustment.